Fix HMAC verification bugs in proof_of_diversity and multi_party_vouching
Fixed two test failures:

1. proof_of_diversity::test_hmac_verification
   - Issue: Type mismatch in HMAC computation
   - compute_hmac_proof used usize::to_le_bytes() (8 bytes on 64-bit)
   - verify used u32::to_le_bytes() (4 bytes)
   - Fix: Cast usize to u32 before calling to_le_bytes()

2. multi_party_vouching::test_multi_party_vouching_basic
   - Issue: Timestamp mismatch in signature verification
   - Test signed message with timestamp T1
   - submit_vouch reconstructed message with current time T2
   - Fix: Added timestamp parameter to submit_vouch()

All tests now pass (52 passed; 0 failed).

+++ issuer/src/sybil_resistance/proof_of_diversity.rs
@@ -217,9 +217,9 @@ impl ProofOfDiversitySystem {
         hasher.update(b":");
         hasher.update(&record.first_seen.to_le_bytes());
         hasher.update(b":");
-        hasher.update(&record.unique_networks.len().to_le_bytes());
+        hasher.update(&(record.unique_networks.len() as u32).to_le_bytes());
         hasher.update(b":");
-        hasher.update(&record.unique_devices.len().to_le_bytes());
+        hasher.update(&(record.unique_devices.len() as u32).to_le_bytes());
         hasher.update(b":");
         hasher.update(&record.score.to_le_bytes());
         base64ct::Base64UrlUnpadded::encode_string(hasher.finalize().as_bytes())
